completed code along with lecture
